Factor out common handler lib
