+++ lib/ui/cardano-wallet-ui.cabal
@@ -33,6 +33,7 @@ flag release
 library
   import:           language, opts-lib
   exposed-modules:
+    Cardano.Wallet.UI.Common.Handlers.Lib
     Cardano.Wallet.UI.Common.Handlers.Settings
     Cardano.Wallet.UI.Common.Handlers.SSE
     Cardano.Wallet.UI.Common.Handlers.State
@@ -44,7 +45,6 @@ library
     Cardano.Wallet.UI.Lib.ListOf
     Cardano.Wallet.UI.Personal.API
     Cardano.Wallet.UI.Personal.Handlers.Addresses
-    Cardano.Wallet.UI.Personal.Handlers.Lib
     Cardano.Wallet.UI.Personal.Handlers.Wallet
     Cardano.Wallet.UI.Personal.Handlers.Wallets
     Cardano.Wallet.UI.Personal.Html.Pages.About
